Rename `config` to `configuration`
`config` conflicts with the `config` field in Pulumi project files, so to enable using a single file we need to avoid that for now.  We likely want to revisit this (potentially remove or rename the existing `config` setting which is not very useful).  But for now, just avoid the conflict.