feat: Deduplicate plugin listings
This also adds diagnostic errors when a single provider is provided with
mutually incompatible version information.

Technically, that's not an issue for our engine, but it could be
confusing in YAML and we'll prevent it for now. It seems more likely
that it's user error if two different versions are specified.

We'll use this de-duplicated plugin list to determine which plugins to
host and get schema from.