Differentiate maps and named types via schema
We need to do differentiate for nested nested types, as well as
functions. To accomplish this, I will extend `importExpr` to carry an
optional type suggestion. Resources and functions will act as the root
for these type hints. When the type suggestion matches the outputted
expression, we will conform the type suggestion. When type suggestions
don't match or are not available, behavior will remain unchanged.

- [  ] Nested types
- [  ] Functions

The initial commit only operates for a single layer under a resource,
and does nothing for functions.