Check schema for maps vs structs (#129)
* Differentiate maps and named types via schema

We need to do differentiate for nested nested types, as well as
functions. To accomplish this, I will extend `importExpr` to carry an
optional type suggestion. Resources and functions will act as the root
for these type hints. When the type suggestion matches the outputted
expression, we will conform the type suggestion. When type suggestions
don't match or are not available, behavior will remain unchanged.

- [X] Nested types
- [X] Functions

The initial commit only operates for a single layer under a resource,
and does nothing for functions.

* Provide deep type hints

The azure-app-service example tests this functionality

* Provide type hints for functions

* Enforce critical section for generate function

* Give node compiler more heap

Tests were failing with OOM when trying to type check ts. My suspicion
is since some of our examples use multiple providers, the type-checking
surface area becomes very memory intensive.