Depend on pu/pu v3.48.0 (#424)
* Upgrade pu/pu dep to v3.48.0

* Remove environmentType from azure-container-apps

This property is not present in the azure-native schema, not does it
pass type check in Pulumi YAML.

* Mock out RunPlugin for *uamlLanguageHost

This is needed for interface compliance.

* Skip typechecking for remaining examples

`azure-app-service` should type check, pulumi/pulumi#11572 will address
the issue.

I'm not sure why `aws-static-website` doesn't type check, the error
message is inscrutable. I opened pulumi/pulumi#11573 to address the
error message. We can come back to the underlying error after the error
message describes the error in a readable way.

* Update TestGenerateProgram tests