Add support for the `deletedWith` resource option
This adds support for the `deletedWith` resource option that has recently been added to the Node.js, Python, .NET, and Go SDKs.

In many cases there is no need to deleted resources if the container resource is going to be deleted as well.

A few examples:
- Database object (roles, tables) when database is being deleted
- Cloud IAM bindings when user itself is being deleted

This helps with:
- Speeding the deletion process
- Removing unnecessary calls to providers
- Avoiding failed deletions when the pulumi user running the plan has access to the container resource but not the contained ones

To avoid deleting contained resources, set the `deletedWith` resource option to the container resource.