ci: Cache downloaded schemas between runs
Our tests download schemas anew each run.
This changes the CI job to cache the downloaded schemas
in the GitHub Actions cache.

A hash of get_schemas.sh is used as the cache key
since that's where we list all required schemas.

To faciliate this, get_schemas.sh gets a new LIST_ONLY option.
When set, it'll only list the files (relative to repository root)
that would be downloaded.
This is passed as input to the actions/cache action,
to let it know which files it should persist in the cache.