go.mod: Tidy with Go 1.19 (#475)
Runs `go mod tidy` with -go=1.19 on the go.mod for the project.
Note that as of Go 1.20, the `go` directive is **not**
the minimum required Go version.
It's the version of the toolchain we developed against.
This is changing in Go 1.21.

The `-compat` flag specifies which version of Go the go.mod file
needs to be compatible with.
By default, this is one release behind the `-go` flag
(or `go` directive).
So we don't need to specify it manually.

Lastly, this deletes the `go mod tidy` in the test job.
Only the lint job should bbe modifying the go.mod
to ensure that the go.mod is up to date.

(This is a copy of #471, which was accidentally merged into #474.)