lint(codegen): Delete unused slices
Fixes the following lint issues by deleting the unused slices.

```
pkg/pulumiyaml/codegen/gen_program.go:230:24: SA4010: this result of append is never used, except maybe in other appends (staticcheck)
pkg/pulumiyaml/codegen/load.go:569:19: SA4010: this result of append is never used, except maybe in other appends (staticcheck)
```

This commit is kept separate from the other lint fixes
because we need someone with more familiarity with this code
to explicitly confirm that this is not a bug in pulumi-yaml.