include name from original project if the compiler doesn't return it (#610)
The pulumi-yaml runtime expects the compiler to return a complete yaml
project, including the `name` attribute. The compiler might not return
this however, even when it returns a valid yaml program otherwise.

This leads to issues down the line, as the rest of the runtime expects
the name to be set in the template, and might panic if it isn't.  We
do however have the name from the original Pulumi.yaml, so we can fill
that in for the user, even if the compiler doesn't return it in the
template.