Test and fix support for local plugins
This updates the pulumi dependency to
[d9df4771815894ab6ddaf17b78f6f023c91b7148](https://github.com/pulumi/pulumi/commit/d9df4771815894ab6ddaf17b78f6f023c91b7148)
which has a number of fixes for plugin and schema loading.

This allows YAML to correctly load schemas for plugins defined in the
Pulumi.yaml file itself. We often use this for integegration tests, but we
expect users to increasinly make use of this feature in the future as an
alternative to dynamic providers.