Allow missing nodes in template (#617)
When a node can not be found, add a missingNode as a placeholder. When
walking the graph, VisitMissing can customize the behaviour for these
nodes. For typechecking and program evaluation missing nodes are always
an error.

During `GetRequiredPlugins` we do not have the program's config
available to us. Previously we would abort the graph walk as soon as we
hit a missing configuration reference. With this PR, we ignore missing
nodes and continue the walk. This allows us detect plugins configured on
resources.