Fix loader lifetime (#640)
Noticed this while working on conformance testing. In conformance
testing we use the grpc loader and once you call Close on that it can't
load schemas anymore. I was seeing tests fail because of Apply calls
that were running after RunTemplate had returned and wanting to do type
checking and loading schemas.

This happened to work for the current plugin loader because it's Close
just clears out caches but it can still load things afterwards, but this
is the more correct place for it.

Firstly once we add conformance loader it can be in the same place,
secondly if plugin host ever does start doing a hard Close we won't
break YAML.