return correct element types for latebound resources (#664)
We currently return a pointer to the interface for latebound resources
from their ElementType methods. However pulumi expects pointers to the
concrete type here, so it can create it and assign values to the created
object. Fix that.

In particular this can happen when calling `fn::secret` on a created
resource. Currently that will panic because the internals get confused
by this wrong `ElementType`. I'm not sure it makes all that much sense
to even call `fn::secret` on this, but we shouldn't panic either.

Fixes https://github.com/pulumi/pulumi/issues/17530. I'm not sure this
completely fixes it as I've managed to produce a slightly different
panic, but not the same one. For lack of better ideas I'm gonna claim
this fixes it for now, and we can revisit if we get any more panic
reports with a pulumi version with this fix included.