Enable l1-builtin-info test (#673)
This enables the l1-builtin-info test. As we're now running multiple
templates with the same `yamlLanguageHost` we can't globally cache the
template in loadTemplate anymore.

Instead we cache by the program directory. For normal runs there will
only be the one program directory, and for conformance test we know each
program will be in its own unique directory.