Enable l1-output-string (#675)
This required two fixes.

Firstly when reading in a literal string from PCL we need to re-escape
any `$${` escape sequences.

Secondly it's valid to have an empty string as a stack export. We fix
this by making sure empty strings from PCL write out as `""` rather than
nothing, as the latter caused the yaml runtime to think it was an
empty/null/missing key rather than an empty string.