Enable l2-resource-simple (#676)
This enables the l2-resource-simple test.

This required a few fixes. 
Firstly in `GetReferencedPlugins` we look at any package declarations to
get package versions. This is so we can return the plugin info with the
right version.

Secondly we add `GetReferencePackages` method to use from
`GetProgramDependencies` so that we can actually return something from
`GetProgramDependencies` rather than nothing. Conformance testing tests
this so we have to fill it in.

Thirdly we implement `Pack`, it doesn't do much just copies the
generated package declaration yaml file.