Enable l2-resource-asset-archive test (#684)
This is a bit more of an involved change to get this test working.
Making two major changes.

Firstly, historically YAML would always send absolute paths instead of
relative paths to the engine. This was wrong, the engine expects that
paths should be relative so that we don't save absolute paths to the
state file.

Secondly, YAML would error if you tried to access what it considered
"unsafe" paths. That was either paths that clearly accessed outside the
Pulumi.yaml directory, or non-constant paths to the remote asset and
archive types. This seems overly restrictive, none of our other
languages try to protect the user like this and it seems reasonable to
expect to be able to do similar filesystem accesses from YAML.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>