Add a Mise setup (#719)
While I'm installing everything, I thought I'd try to make things
consistent.
This mirrors the `pulumi/pulumi` PR here:
https://github.com/pulumi/pulumi/pull/18626/

This commit adds a quick intro for `mise` as a tool for building the
repository, and a minimal `.mise.toml` to include all the dependencies
required to run the `make` commands.