factor out a `Template` interface and use it in the runner
In subsequent PRs we're going to introduce components into yaml.  A
YAML program will be able to have multiple components, that will be
run similar to programs, but won't have all the parts of programs.
E.g. a YAML template will be able to include components, but
components won't be able to have further components.  Or components
won't have packages just for the component itself (but will respect
the global packages).

To be able to reuse the runner for running components, factor out a
template interface and use it.

No functional changes intended.