gofumpt yaml repository
`pulumi/pulumi` requires all files to be formatted with `gofumpt`, and
enforces that with a linter.  Do the same for the YAML repository for
consistency.

Not having this `gofumpt`ed is especially annoying as I have an editor
extension that does just that on every save, so I'd need to treat this
repo specially in my editor.  Having it `gofumpt`ed seems like a good
change either way, so do that, and make my life a bit easier for
further changes that are going to be introduced.