run tests without -v
Running tests with -v means we show all the output independently of
whether a test succeeded or not.  Some of the conformance tests have
extremely long outputs, even when the test succeeds, so when there are
any failures GitHub actions can't handle the amount of logs, and
doesn't show the actuall error anymore.  Fix that by removing the `-v`
from tests.  Failed tests will still show the output correctly, and we
don't need all the output for successful tests.