run tests without -v
Running tests with -v means we show all the output independently of
whether a test succeeded or not.  This makes the test output in CI
much harder to sort through, as there is a ton of text output for
successful text that is not very useful, and the output for failed
tests gets hard to find.