log only part of stdout in conformance tests
The output of these conformance tests can become huge in some cases.
We always log the whole stdout and stderr here, which includes all the
outputs for conformance tests.

In this repo we run the tests with `-v`, so if any test fails, we
print the output of all conformance tests in full.  This can become so
big that GitHub actions can no longer handle the text, and just gives
a generic "Error: " message, which is useless for debugging what's
happening, or even seeing which test failed.

Similar to how we already clamp down the logs sent to the engine, also
clamp down this log, with the option of showing the full log if
desired.

Arguably we could run the tests without passing `-v` to `go test`, but
that just means that if a conformance test with such a long output
fails we end up in the same position again, struggling to debug.

We should probably implement the same in other languages as well, but
it's less noticable there because we don't run the tests in
`pulumi/pulumi` with `-v`.