remove PULUMI_STACK from test to check if we keep conflicting values (#742)
Since https://github.com/pulumi/pulumi/pull/18717, PULUMI_STACK is used
to select the stack we're using. Therefore it is now impossible for the
env var to conflict with the stack we're using. Just remove the test for
`PULUMI_STACK` from this test, as it's impossible to make it conflict
with the stack that's actually going to be used, which is what the test
is testing.

There are other tests that check that `PULUMI_STACK` is set correctly.

Fixes https://github.com/pulumi/pulumi-yaml/issues/738