implement schema generation for yaml components
For components to work we need to generate their schema.  In YAML we
currently only support strings, integers, booleans and arrays as types
for config declarations, so let's support the same thing for YAML
components for now.

Outputs types are not super straightforward to infer, so right now
we're sticking to `pulumi.json/#Any`.  In future revisions we can
support more accurate types here.