Rename packages AST to sdks (#757)
Apparently, we use a fake `Packages` property in template's AST to hold
metadata from generated SDKs (package descriptions). It's not a real AST
node but we populate it manually to bring those SDKs into the same
context.

Now, that we've introduced a proper `packages` section in Pulumi.yaml,
this suddenly started to panic. This PR is a minimal non-functional
change to avoid this problem by renaming the property and related
methods to `Sdks`/`GetSdks`, which is closer to our common nomenclature.

The `packages` section support will come in a follow-up PR.

Fix https://github.com/pulumi/pulumi-yaml/issues/756