Support the packages Pulumi.yaml section in YAML runtime (#758)
Update to the latest pulumi/pulumi version that added support for the
`packages` section in Pulumi.yaml. We need to thread the packages
definition into the YAML loader to make sure it works for the YAML
runtime as well.

To test the change, I changed the existing component integration test to
replace the `plugins` section with `packages`.

Part of https://github.com/pulumi/pulumi/issues/18918