allow config defaults to be integers
In YAML it is possible to set a default value for a config entry.
However YAML converts all number looking things into `float64`s/`number`s,
which makes the config system reject an integer default value for an
integer type.

Since there isn't a great way to distinguish floats and ints in YAML,
let the config system continue to parse all numbers into floats.
However when applying the default value, we check if it can fit an
int, and convert it to an int in that case.