pass through pluginDownloadURL when loading plugin
pulumi-yaml supports a `pluginDownloadURL` resource option.  However
when trying to load the package that URL is not passed through
currently, and thus is not used when trying to get the package.

We should always be using it if it is available, and let the loading
code handle what it wants to do with it.  Do that here.