pass through pluginDownloadURL when loading plugin (#776)
pulumi-yaml supports a `pluginDownloadURL` resource option. However when
trying to load the package that URL is not passed through currently, and
thus is not used when trying to get the package.

We should always be using it if it is available, and let the loading
code handle what it wants to do with it. Do that here.

This isn't meant to be the main way to load components anymore, but we
should still deal with the option correctly.

Fixes https://github.com/pulumi/pulumi/issues/18638