Allow resource properties to refer to an object variable  (#779)
Fixes https://github.com/pulumi/pulumi-yaml/issues/770

This allows setting the properties of a resource with an object
expression rather a fixed object:
```
resources:
  randomPassword:
    type: random:RandomPassword
    properties: ${inputs}
```
Where for example inputs is a config, object variable, or possible the
output of another resource.

This is done by allowing either a `PropertyMapDecl` or an `Expr` for the
"properties" key. For the `PropertyMap` case nothing changes. When we
have an `Expr` we type check that the expression type matches the object
shape of the resource. To actually set the properties depends on if the
expression resolves as a plain map, in which case we simply copy each
value off the map. Or if it resolves as an output, in which case we
generate an Apply for each expected input property (given by the
resource schema) to copy _that_ field from the map we expect the output
to eventually resolve to.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>