Downgrade to go 1.23 (#798)
Unfortunately pulumi-terraform-bridge library build-depends on
pulumi-yaml as a library and therefore go 1.24 language requirement is a
bit more infectious than anticipated and needs to affect all downstream
bridged providers. Since pulumi/pulumi is itself on go 1.23 for the
moment, this downgrades to 1.23 to match; the idea is to upgrade
everything in Pulumi to 1.24 when the time comes to do that.

The HELPMAKEGO utility was automatically upgrading the project to go1.24
this is now fixed with `go install -mod readonly`.

https://github.com/pulumi/pulumi-terraform-bridge/issues/3077 reminder
to chase down and remove the build-dependency.