Fix LoadTemplate resolving the project directory to . on Windows (#801)
In https://github.com/pulumi/pulumi/pull/19582 we changed the plugin
context of the convert process to be the output directory where the
generated project lives. The change was simple but suddenly a smoke test
started to fail on _Windows_ and it wasn't a flake:
```
=== FAIL: smoke TestYamlConvertSmoke (3.19s)
    environment.go:88: Created new test environment:  C:\Users\RUNNER~1\AppData\Local\Temp\test-env3425171743
    environment.go:197: Running command D:\a\pulumi\pulumi\bin\pulumi plugin install resource random 4.13.0
    environment.go:197: Running command D:\a\pulumi\pulumi\bin\pulumi convert --strict --language pcl --from yaml --out out
    smoke_test.go:[96](https://github.com/pulumi/pulumi/actions/runs/15116997853/job/42498541044?pr=19582#step:41:97): Run Error: exit status 0xffffffff
    smoke_test.go:96: STDOUT: 
    smoke_test.go:96: STDERR: Converting from yaml...
        Downloading provider: yaml
        error: rpc error: code = Unknown desc = load yaml program: reading template Main: open Pulumi.yaml: The system cannot find the file specified.
    smoke_test.go:96: Ran command pulumi args [convert --strict --language pcl --from yaml --out out] and expected success. Instead got failure.
```
Even though the `ConvertProgramRequest` hasn't changed at all and
pulumi-converter-yaml does use the `SourceDirectory` correctly.

After hours of debugging on my personal Windows machine and figuring out
how to build everything, I was able to track it down to
`LoadTemplate(dir string)` resolving the project directory using
`projectDir := path.Dir(projectPath)` where `path.Dir(...)` returned `.`
(implicit working directory) instead of using the the dir which is the
in the input of the function.

This problem only now surfaced because before
https://github.com/pulumi/pulumi/pull/19582, the implicit working
directory was the YAML project directory but now it is the output
directory (where PCL is generated)

This PR fixes the problem by not using `path.Dir`, instead using
`filepath.Dir`