Revert back to string based config (#807)
Tried to just do a revert of 07c615b9370ebafbc08af2ce19e71aa9a8657e14,
but with a few extra fixes in `getConfigNode` to less eagerly convert
values like "01234" into the number "1234". But hit issues where bits of
the system had diverged to depending on the property value info. So
changed to just rebuild the property value map at startup from the
textual config.

Probably not the long term ideal fix, but it stops us from looking at
`RunRequest.ConfigPropertyMap` so we can take that out the engine.

This aligns yaml with the other runtimes that the engine doesn't make
config typing decisions. If users need extra features to be able to
treat `"true"` as the string instead of the boolean we can add a
`getRawConfigText` function or the like.