show diagnostics correctly when generating program returns them
In `pulumi-convert-yaml`, we try to generate a PCL program from the
YAML template.  This can return diagnostics if there are any issues in
the program, and we fail to generate the program correctly, or have
any warnings.

However when checking whether we should show these diagnostics to the
user, we check the `diags` variable instead of `pdiags`.  The `diags`
variable comes from checking the template, not from trying to generate
the program.  So we often end up not showing any diagnostics, even
though we should.  Fix this by checking the right variable.