Add `object` as a type in `pulumi-yaml` (#848)
Currently, we run into issues with the latest release because we have no way, as users, to specify that some config variable is an opaque `object`. This causes problems with, for example, the pattern of creating a `props` variable to pass props to other parts of the structure. This PR introduces the `object` type - it's just a `map[string]any` - and allows users to specify it as a type in config(uration). With this, we'll be able to close out #832.