Include `config` in type-checking (#861)
In #827, a bug was reported because we started returning an error that
we previously hadn't wired up. This is good because it meant that we
caught more bad programs, but bad because it exposed a few bugs in the
code that previously weren't visible.

The bug in question happened because we only ever included
`configuration` and never `config` in our analysis. The knock-on result
is that any variables declared within `config` would be considered
undefined elsewhere. This PR simply updates the `GetConfig` method
(which previously returned `configuration`) to return `config` and
anything in `configuration` whose key wasn't in `config`. In other
words, we now return `config` and `configuration` but give priority to
`config`.

Another result of this change is that now we have to declare types for
variables in `config`: no longer will `props: {}` be accepted because it
contains neither a type nor a default value. I've updated the unhappy
tests to include both, and all other tests seem to work happily with
config maps as before.

I'm a little unsure about this change, so I'm hoping someone can help to
validate it. I've spent a _long_ time going round in circles with this
trying to understand whether what I'm looking at is a bigger bug (i.e.
there's something fundamentally wrong with the way we handle config vs
configuration, there's something wrong with the code path for `pulumi
convert`), but I think ultimately this fixes the problem with a diff
about a thousand lines smaller. If anyone here is a `pulumi-yaml`
expert, feedback is very much appreciated.