Support scalar invokes (#885)
YAML support for https://github.com/pulumi/pulumi/issues/7435

This checks package schemas to see if functions are declared with
`ReturnType` rather than `Outputs`, and if so to pick off the single
value from the returned property map.

We also error (maybe should just be a warning) that you can't use
`return` with an invoke returning a single value.