Revert the upgrade to the latest Pulumi and use a different commit (#904)
We needed to upgrade Pulumi, but the upgrade we made included
https://github.com/pulumi/pulumi/commit/66146e67a2490af257f50b9a13b5b6506d74caed,
which upgrades `github.com/iancoleman/strcase` from v0.2.0 to v0.3.0.
Unfortunately, this has a breaking behavior change.

We are going to revert the pu/pu bump, but in the meantime, we can get a
release out of `pulumi-yaml` faster by reverting the original upgrade
commit, and then upgrading to the commit just before
https://github.com/pulumi/pulumi/commit/66146e67a2490af257f50b9a13b5b6506d74caed.

Ref: https://github.com/pulumi/pulumi/issues/21217